searchengine/python/requirements.txt
# 向量计算相关
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.12.0

# 向量索引相关
faiss-cpu>=1.7.0

# 数据库相关
mysql-connector-python>=8.0.0
pymysql>=1.0.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 进度条
tqdm>=4.64.0